feat 视频转码

+++ browse/browse-client/src/main/java/org/ht/model/dto/VideoMetadataDTO.java
@@ -82,7 +82,7 @@ public class VideoMetadataDTO {
     /**
      * 创建者ID
      */
-    private Long creatorId;
+    private Integer creatorId;
 
     /**
      * 分类id
